Translation 1
The first part of my attempt at translating this addon. There's still quite a bit left to deal with and the reason it takes time is because in order to properly translate it (not just translate but have the text actually make sense to the player) I need to use the features to know what they do so I know what the text should say. That doesn't work when half the buttons do nothing though. There is a lot of text I can't clean up because I can't see it in-game to know what it relates to.

This is just the first part of more to come. Just making sure the work I've done so far is not lost.